requirements.txt
streamlit==1.38.0
SQLAlchemy==2.0.36
pydantic==2.9.2
python-dotenv==1.0.1
APScheduler==3.10.4
pandas==2.2.2
web3==6.20.1
requests==2.32.3
